requirements.txt
python-rtmidi==1.5.8
pillow~=11.1.0